refactoring previous engine.py to main.py engine logic move to engine.py class
